\name{MPBoost-package}

\alias{MPBoost-package}

\alias{MPBoost}

\docType{package}

\title{\packageTitle{MPBoost}}

\description{\packageDescription{MPBoost}}

\details{\packageIndices{MPBoost}}

\author{
  Ignacio López-de-Ullibarri
  
  Maintainer: Ignacio López-de-Ullibarri <ignacio.lopezdeullibarri@udc.es>
}

\references{Berger, V. W., Ivanova, A., Knoll, M. D. (2003). Minimizing
  predictability while retaining balance through the use of less
  restrictive randomization procedures. Statistics in Medicine, 22: 3017-3028.
  \url{https://doi.org/10.1002/sim.1538}.
  
  Salama, I., Ivanova, A., Qaqish, B. (2008). Efficient generation of
  constrained block allocation sequences. Statistics in Medicine, 27,
  1421-1428. \url{https://doi.org/10.1002/sim.3014}. 
}

\keyword{package}

